Inventory Done
The inventory system now works. All that is really left to do for it is to be able to use items, and finding new items to pickup and add it to the inventory in the correct list. Will work on tweaking it a bit and then start adding Pemo tamers.